Add files via upload
Add support for blank "raFiles" parameter, which implies that we'll compute the required verses to fetch for recordings directly in this script rather than inheriting it from scrollscraper.cgi.

Also add initial support for 'httpStyle' parameter which if true indicates that we want to provide REST-like responses for 302, 202 and 503 HTTP scenarios.

Where:
302 = redirect for a result we've already computed
503 = you've hit limits for today
202 = we're computing the result now.    Try again soon and we'll have them completed.

But this logic isn't quite right yet.    Among other things the HTTP headers aren't right.    We also might need a daemon to run the computational shell script since I don't know how to persuade nginx to return a partial result when a child process (the computational shell script) is still running.